move map functions to user_functions
