--- py/sentry_streams/sample_function.py
@@ -1,15 +1,0 @@
-import json
-
-
-class EventsPipelineMapFunction:
-    """
-    Sample user-defined functions to
-    plug into pipeline
-    """
-
-    @staticmethod
-    def simple_map(value: str) -> str:
-        d = json.loads(value)
-        res: str = d.get("name", "no name")
-
-        return "hello " + res
