+++ py/sentry_streams/user_functions/sample_filter.py
@@ -14,3 +14,17 @@ class EventsPipelineFilterFunctions:
         Filter function with wrong return type, used in tests
         """
         pass
+
+
+class EventsPipelineMapFunctions:
+    """
+    Sample user-defined functions to
+    plug into pipeline
+    """
+
+    @staticmethod
+    def simple_map(value: str) -> str:
+        d = json.loads(value)
+        res: str = d.get("name", "no name")
+
+        return "hello " + res
