+++ py/tests/test_pipeline.py
@@ -9,8 +9,10 @@ from sentry_streams.adapters.stream_adapter import RuntimeTranslator, StreamAdap
 from sentry_streams.flink.flink_adapter import FlinkAdapter
 from sentry_streams.pipeline import Filter, KafkaSink, KafkaSource, Map, Pipeline
 from sentry_streams.runner import iterate_edges
-from sentry_streams.sample_function import EventsPipelineMapFunction
-from sentry_streams.user_functions.sample_filter import EventsPipelineFilterFunctions
+from sentry_streams.user_functions.sample_filter import (
+    EventsPipelineFilterFunctions,
+    EventsPipelineMapFunctions,
+)
 
 
 @pytest.fixture(autouse=True)
@@ -104,7 +106,7 @@ def basic_map() -> tuple[Pipeline, MutableMapping[str, list[dict[str, Any]]]]:
         name="mymap",
         ctx=pipeline,
         inputs=[source],
-        function=EventsPipelineMapFunction.simple_map,
+        function=EventsPipelineMapFunctions.simple_map,
     )
 
     _ = KafkaSink(
